tls/server: move server.go in its proper location
